+++ Source/Serilog.Exceptions/Core/ExceptionPropertiesBag.cs
@@ -2,6 +2,7 @@ namespace Serilog.Exceptions.Core
 {
     using System;
     using System.Collections.Generic;
+    using Serilog.Exceptions.Filters;
 
     internal class ExceptionPropertiesBag : IExceptionPropertiesBag
     {
